connection sortProducts => recipes
