+++ .idea/deploymentTargetDropDown.xml
@@ -7,11 +7,11 @@
         <deviceKey>
           <Key>
             <type value="VIRTUAL_DEVICE_PATH" />
-            <value value="C:\Users\97250\.android\avd\Pixel_4_API_30.avd" />
+            <value value="C:\Users\97250\.android\avd\Pixel_5_API_31.avd" />
           </Key>
         </deviceKey>
       </Target>
     </targetSelectedWithDropDown>
-    <timeTargetWasSelectedWithDropDown value="2022-12-01T20:23:58.332276800Z" />
+    <timeTargetWasSelectedWithDropDown value="2022-12-04T07:33:57.518847600Z" />
   </component>
 </project>
