+++ app/src/main/java/com/example/smartfridge/MainActivity.java
@@ -9,6 +9,9 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.google.android.material.button.MaterialButton;
 
+import costumer.customer_USER;
+import manager.management_USER;
+
 public class MainActivity extends AppCompatActivity {
 
     @Override
