--- app/src/main/java/com/example/smartfridge/page_dryFood.java
@@ -1,14 +1,0 @@
-package com.example.smartfridge;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-
-public class page_dryFood extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_page_dry_food);
-    }
-}
