+++ app/src/main/java/costumer/list_item.java
@@ -1,12 +1,13 @@
-package com.example.smartfridge;
+package costumer;
 
 import android.os.Bundle;
-
-import androidx.fragment.app.Fragment;
-
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.smartfridge.R;
 
 /**
  * A simple {@link Fragment} subclass.
