+++ app/src/main/java/costumer/page_dryFood.java
@@ -1,14 +1,16 @@
-package com.example.smartfridge;
+package costumer;
 
 import androidx.appcompat.app.AppCompatActivity;
 
 import android.os.Bundle;
 
-public class page_milky extends AppCompatActivity {
+import com.example.smartfridge.R;
+
+public class page_dryFood extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_page_milky);
+        setContentView(R.layout.activity_page_dry_food);
     }
 }
