+++ app/src/main/java/costumer/page_meat.java
@@ -1,6 +1,4 @@
-package com.example.smartfridge;
-
-import androidx.appcompat.app.AppCompatActivity;
+package costumer;
 
 import android.annotation.SuppressLint;
 import android.content.SharedPreferences;
@@ -10,8 +8,12 @@ import android.widget.Button;
 import android.widget.EditText;
 import android.widget.TextView;
 
-import com.google.gson.reflect.TypeToken;
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.smartfridge.ModelClass;
+import com.example.smartfridge.R;
 import com.google.gson.Gson;
+import com.google.gson.reflect.TypeToken;
 
 import java.lang.reflect.Type;
 import java.util.ArrayList;
