+++ app/src/main/java/costumer/page_milky.java
@@ -1,0 +1,16 @@
+package costumer;
+
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.example.smartfridge.R;
+
+public class page_milky extends AppCompatActivity {
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_page_milky);
+    }
+}
