+++ app/src/main/java/manager/management_USER.java
@@ -1,4 +1,4 @@
-package com.example.smartfridge;
+package manager;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -8,6 +8,7 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.example.smartfridge.R;
 import com.google.android.material.button.MaterialButton;
 
 public class management_USER extends AppCompatActivity {
