+++ app/src/main/res/layout/activity_management_user.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/background2"
-    tools:context=".customer_USER">
+    tools:context="costumer.customer_USER">
 
     <TextView
         android:layout_width="match_parent"
