+++ app/src/main/res/layout/activity_manager.xml
@@ -8,7 +8,7 @@
     android:foregroundTint="#EC3A3A"
     android:orientation="vertical"
     android:background="@drawable/background2"
-    tools:context=".costumers">
+    tools:context="costumer.costumers">
 
     <LinearLayout
         android:id="@+id/category"
