+++ app/src/main/res/layout/activity_page_dry_food.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".page_dryFood">
+    tools:context="costumer.page_dryFood">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
