+++ app/src/main/res/layout/activity_page_meat.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@drawable/background2"
-    tools:context=".page_meat">
+    tools:context="costumer.page_meat">
 
     <EditText
         android:id="@+id/et_name"
