+++ app/src/main/res/layout/activity_page_vegetables.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".page_vegetables">
+    tools:context="costumer.page_vegetables">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
