+++ app/src/main/res/layout/activity_recipes.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:padding="10dp"
-    tools:context=".recipes">
+    tools:context="costumer.recipes">
 
     <TextView
         android:layout_width="wrap_content"
