+++ app/src/main/res/layout/activity_shopping_tables.xml
@@ -7,7 +7,7 @@
     android:layout_height="match_parent"
     android:orientation="vertical"
     android:background="@drawable/background2"
-    tools:context=".ShoppingTables">
+    tools:context="costumer.ShoppingTables">
 
     <LinearLayout
         android:id="@+id/welcom"
