+++ app/src/main/res/layout/activity_upload_recipes.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Upload_Recipes">
+    tools:context="manager.Upload_Recipes">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
